+++ include/vb/Minimizer.h
@@ -4,9 +4,9 @@
 namespace vb {
     template <typename T> class Minimizer {
     public:
-        Minimizer(int n_);
-        Minimizer(int n_, std::function<T(const Vector<T> &, Vector<T> *)> fg_);
-        Minimizer(int n_, std::function<T(const Vector<T> &)> f_, std::function<Vector<T>(const Vector<T> &)> g_);
+        Minimizer(size_t n_);
+        Minimizer(size_t n_, std::function<T(const Vector<T> &, Vector<T> *)> fg_);
+        Minimizer(size_t n_, std::function<T(const Vector<T> &)> f_, std::function<Vector<T>(const Vector<T> &)> g_);
 
         T compute(const Vector<T> &x = Vector<T>(0));
 
@@ -18,7 +18,7 @@ namespace vb {
         T minimize_pr(const Vector<T> &x0);
         T minimize_qn(const Vector<T> &x0);
 
-        int n = 0;
+        size_t n = 0;
 
         std::function<T(const Vector<T> &)>              f;
         std::function<Vector<T>(const Vector<T> &)>      g;
